ai-brain/requirements.txt
langchain
langgraph
langchain-google-genai
nats-py
psycopg2-binary